app_files/requirements.txt
langchain
streamlit
PyPDF2
pinecone-client
python-dotenv
pypdf
openai
tiktoken
InstructorEmbedding
transformers
datasets
jsonlines
numpy
requests
scikit_learn
scipy
sentence_transformers
torch
tqdm
rich